+++ src/orchestration/individual-member-sdk.ts
@@ -1,8 +1,8 @@
 import type { DataspaceNodeClient } from '../client.js';
-import { GdcProfessionalSdk } from '../../../gdc-sdk-node-ts/dist/index.js';
+import { GdcIndividualMemberSdk } from '../../../gdc-sdk-node-ts/dist/index.js';
 import { DataspaceNodeRuntimeClientAdapter } from '../gdc-node-runtime-client-adapter.js';
 
-export class ProfessionalSdk extends GdcProfessionalSdk {
+export class IndividualMemberSdk extends GdcIndividualMemberSdk {
   constructor(client: DataspaceNodeClient) {
     super(new DataspaceNodeRuntimeClientAdapter(client));
   }