+++ src/orchestration/organization-controller-sdk.ts
@@ -1,8 +1,8 @@
 import type { DataspaceNodeClient } from '../client.js';
-import { GdcProfessionalSdk } from '../../../gdc-sdk-node-ts/dist/index.js';
+import { GdcOrganizationControllerSdk } from '../../../gdc-sdk-node-ts/dist/index.js';
 import { DataspaceNodeRuntimeClientAdapter } from '../gdc-node-runtime-client-adapter.js';
 
-export class ProfessionalSdk extends GdcProfessionalSdk {
+export class OrganizationControllerSdk extends GdcOrganizationControllerSdk {
   constructor(client: DataspaceNodeClient) {
     super(new DataspaceNodeRuntimeClientAdapter(client));
   }